+++ src/main/java/org/altbeacon/beacon/Beacon.java
@@ -285,14 +285,25 @@ public class Beacon implements Parcelable, Serializable {
     }
 
     /**
-     * Returns the running average rssi
-     * @param rssi
-     * @return
+     * @deprecated To be removed in a future release. Use
+     * {@link org.altbeacon.beacon.Beacon#getRunningAverageRssi()}
+     * instead.
      */
+    @Deprecated
     public double getRunningAverageRssi(double rssi) {
         return mRunningAverageRssi = rssi;
     }
 
+    /**
+     * Returns the running average rssi
+     * @return double
+     */
+    public double getRunningAverageRssi() {
+        if (mRunningAverageRssi != null){
+            return mRunningAverageRssi;
+        }
+        return mRssi;
+    }
 
     /**
      * Sets the most recently measured rssi for use in distance calculations if a running average is